feat: impl dagger app base part + hero wars part + global refactor
