--- app/src/main/java/com/example/guillaume_grand_clement/herowars/data/pojo/SamplePojo.java
@@ -1,15 +1,0 @@
-package com.example.guillaume_grand_clement.herowars.data.pojo;
-
-import io.realm.RealmObject;
-
-public class SamplePojo extends RealmObject {
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
