+++ app/src/main/java/com/herowars/template/dagger/base/NetComponent.java
@@ -1,0 +1,20 @@
+package com.herowars.template.dagger.base;
+
+import javax.inject.Singleton;
+
+import dagger.Component;
+import okhttp3.OkHttpClient;
+import retrofit2.Retrofit;
+
+/**
+ * Dans cette classe doit être déclaré explicitement
+ * les références d'objets déclarés dans les modules
+ */
+@Singleton
+@Component(modules = {AppModule.class, NetModule.class})
+public interface NetComponent {
+
+    Retrofit retrofit();
+
+    OkHttpClient okHttpClient();
+}
