+++ app/src/main/java/com/herowars/template/dagger/herowars/HeroWarsComponent.java
@@ -1,0 +1,13 @@
+package com.herowars.template.dagger.herowars;
+
+import com.herowars.template.activity.AbsActivity;
+import com.herowars.template.dagger.base.NetComponent;
+
+import dagger.Component;
+
+@HeroWarsScope
+@Component(dependencies = NetComponent.class, modules = HeroWarsModule.class)
+public interface HeroWarsComponent {
+
+    void inject(AbsActivity absActivity);
+}
