+++ app/src/main/java/com/herowars/template/dagger/herowars/HeroWarsScope.java
@@ -1,0 +1,8 @@
+package com.herowars.template.dagger.herowars;
+
+import javax.inject.Scope;
+
+@Scope
+public @interface HeroWarsScope {
+
+}
