+++ app/src/main/java/com/herowars/template/data/pojo/SamplePojo.java
@@ -1,0 +1,15 @@
+package com.herowars.template.data.pojo;
+
+import io.realm.RealmObject;
+
+public class SamplePojo extends RealmObject {
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
