+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/coordinator_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.activity.impl.MainActivity">
+    tools:context="com.herowars.template.activity.impl.MainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
